refactor(core): 改为单流
