+++ go.mod
@@ -3,6 +3,6 @@ module github.com/yomorun/yomo
 go 1.14
 
 require (
-	github.com/10cella/yomo-txtkv-codec v1.0.5
+	github.com/10cella/yomo-txtkv-codec v1.0.6
 	github.com/lucas-clemente/quic-go v0.17.1
 )
