+++ pkg/yomo/run.go
@@ -7,8 +7,6 @@ import (
 	"os"
 	"time"
 
-	txtkv "github.com/10cella/yomo-txtkv-codec"
-
 	"github.com/yomorun/yomo/pkg/plugin"
 	"github.com/yomorun/yomo/pkg/util"
 
@@ -21,26 +19,8 @@ func Run(plugin plugin.YomoObjectPlugin, endpoint string) {
 	log.SetPrefix(fmt.Sprintf("[%s:%v]", plugin.Name(), os.Getpid()))
 	log.Printf("plugin service start... [%s]", endpoint)
 
-	// binding plugin
-	pluginStream := framework.NewObjectPlugin(plugin)
-
-	// decoding
-	deStream1 := txtkv.NewObjectDecoder(plugin.Observed())
-
-	//过滤
-	deStream2 := txtkv.NewFilterDecoder(plugin.Observed())
-
-	// encoding
-	enStream := txtkv.NewObjectEncoder(plugin.Observed())
-
-	deStream := io.MultiWriter(deStream1.Writer, deStream2.Writer)
-
-	go func() { io.CopyN(pluginStream.Writer, deStream1.Reader, 1024) }() // nolint
-	go func() { io.CopyN(enStream.Writer, pluginStream.Reader, 1024) }()  // nolint
-	go func() { io.CopyN(enStream.Writer, deStream2.Reader, 1024) }()     // nolint
-
 	// activation service
-	framework.NewServer(endpoint, deStream, enStream.Reader)
+	framework.NewServer(endpoint, plugin)
 }
 
 // RunStream run a server for YomoStreamPlugin
@@ -48,26 +28,8 @@ func RunStream(plugin plugin.YomoStreamPlugin, endpoint string) {
 	log.SetPrefix(fmt.Sprintf("[%s:%v]", plugin.Name(), os.Getpid()))
 	log.Printf("plugin service start... [%s]", endpoint)
 
-	// binding plugin
-	pluginStream := framework.NewStreamPlugin(plugin)
-
-	// decoding
-	deStream1 := txtkv.NewStreamDecoder(plugin.Observed())
-
-	//过滤
-	deStream2 := txtkv.NewFilterDecoder(plugin.Observed())
-
-	// encoding
-	enStream := txtkv.NewStreamEncoder(plugin.Observed())
-
-	deStream := io.MultiWriter(deStream1.Writer, deStream2.Writer)
-
 	// activation service
-	framework.NewServer(endpoint, deStream, enStream.Reader)
-
-	go func() { io.CopyN(pluginStream.Writer, deStream1.Reader, 1024) }() // nolint
-	go func() { io.CopyN(enStream.Writer, pluginStream.Reader, 1024) }()  // nolint
-	go func() { io.CopyN(enStream.Writer, deStream2.Reader, 1024) }()     // nolint
+	panic("not impl")
 }
 
 // RunDev makes test plugin connect to a demo YoMo server
@@ -77,26 +39,8 @@ func RunDev(plugin plugin.YomoObjectPlugin, endpoint string) {
 		log.SetPrefix(fmt.Sprintf("[%s:%v]", plugin.Name(), os.Getpid()))
 		log.Printf("plugin service start... [%s]", endpoint)
 
-		// binding plugin
-		pluginStream := framework.NewObjectPlugin(plugin)
-
-		// decoding
-		deStream1 := txtkv.NewObjectDecoder(plugin.Observed())
-
-		//过滤
-		deStream2 := txtkv.NewFilterDecoder(plugin.Observed())
-
-		// encoding
-		enStream := txtkv.NewObjectEncoder(plugin.Observed())
-
-		deStream := io.MultiWriter(deStream1.Writer, deStream2.Writer)
-
-		go func() { io.CopyN(pluginStream.Writer, deStream1.Reader, 1024) }() // nolint
-		go func() { io.CopyN(enStream.Writer, pluginStream.Reader, 1024) }()  // nolint
-		go func() { io.CopyN(enStream.Writer, deStream2.Reader, 1024) }()     // nolint
-
 		// activation service
-		framework.NewServer(endpoint, deStream, enStream.Reader)
+		framework.NewServer(endpoint, plugin)
 	}()
 
 	yomoEchoClient, err := util.QuicClient("echo.cella.fun:11521")
